Add support for auth socket path
OpenSSH uses the SSH_AUTH_SOCK environment variable to
determine the path of the unix domain socket for the agent.

libssh2 has an API that lets us set it, but it was not
exposed by NMSSH.

This commit adds an API for it to NMSSHSession which you can
set prior to calling -connectToAgent.